Add support for high DPI screens
There's a ratio between the pixels we develop the web for and the real
amount of pixels that the screen is using: window.devicePixelRatio.
Normal screens have a ratio of 1, but high DPI screens may have 2.5 or
3. This ratio is applied by the browser to the html elements, but not to
the content of the canvas.

To fix this we use this ratio to increase the amount of pixels that
contains the canvas  but we keep its width and height through css. This
produces that the canvas has a higher pixel density.

To work properly it needs the spice server to be prepared with bigger
resolution and a bigger UI.