Rate-limit 'wheel' events
When smooth scrolling is enabled (touch screens, high-end touchpads),
the browser emits a lot of wheel events with very small deltas.

We aren't sending those deltas to the server, so we ended up scrolling
way too fast on those cases.

This commit limits the amount of 'wheel' events sent to the server.