Added disableMessageBuffering to disable message buffering
Buffering is an optimization technique used in the spice web client to
optimize video processing on slow computers.
In order to avoid processing unnecesary frames, packets are processed at
batches of 50ms, allowing to create a queue of packets and to detect
duplicated messages that are completly overlaping.
It makes no sense to process a message to draw an image that will be
present in the scene for less than 50ms.
However, this creates lag and a feeling of slowness under normal
cirscumstances. A new setting disableMessageBuffering allows to
enable/disable this feature.
The default value is false, so buffering is enabled by default, for
backward compatibility with old settings.
However, for normal scenarios it is recommended to set this setting to
true, improving the overall experience at the expense of decreasing the
performance when playing videos through spice.