Removed unnecesary getImageFromData when CLIENT_RECTS
When adding swcanvas support, an innecesary getImageFromData was added
to drawImage because of lack for clipping support in swcanvas. However,
swcanvas is slower than canvas from almost all the operations, so this
support is unnecesary. swcanvas is still in the repos for historical
reasons and as a possible future work if swcanvas is faster in some
environment.