spice cache activated again, limited to ~150mb
SPICE cache is a cache of bitmaps stored client side but controlled
server side. The server tells the client which bitmaps to store for
future use and which not. During the handshaking on the display channel
a value indicating your desired cache size is sent to the server. This
value is supposed to be the ammount of memory you want to use for cache.
However, being in a browser with all the abstractions for data
structures, there is a very big difference between how much memory we
ask to consume for cache and how much its really used.
Doing some trial and error I found that with the current configuration
the client consumes around 150mb of cache after extensive use.

More testing is needed to better understand the cache.