Improved drawImage performance by avoding to create subrectangles
rasterEngine->drawImage always created new images when a subrectangle
was requested. drawImage and putImageData already allows to use
subrectangles when drawing from an ImageData or a Canvas, so creating a
subrectangle and then drawing to the target can be simplified to one
simple draw call of a subarea. This is imposible if a ROP other than PUT
is requested or a scaled version is requested this is not possible
because we need the subrectangle to do the math.

In 99% of the cases the operation is put and a simple subrectangle is
requested with no scaling, so a direct putimagedata/drawimage can be
used saving one draw call for operation.