Emit error when handshake is longer than 5 seconds
We consider handshake the time between the start of the connection and
the moment we receive the first spice packet.

It should take a few miliseconds (depending on the connection), so if it
takes 5 seconds, we need to be able to recover from this and be able to
do a reconnect or whatever. That's why we emit an error with 4200.

(4200 is a custom code for remote machine lost)