Avoid tab key to get stuck when pressing Ctrl+tab
In firefox, when pressing ctrl+tab, changes the browsers tab but we the
key event is sent. Since the keyup is done outside the document, the tab
key gets stuck and a lot of tabs appear where you were writting.

Using the visibility api, we check when the document is hidden and send
a keyup event for all keys.