Proper treatment of shift key
In this commit we enable again sending shift keys, so, we now will be
able to select text with shift and make all shift combinations.

Fixes bugs like this:
    - having keyboard in US layout
    - pressing shift + 2 to produce "@" sign
    - we send "shift + altgr + 2" producing a weird sign instead of "@".

Basically we control if there's a shift pressed and if the shortcut
contains a shiftdown, we don't send it. In case it's pressed, we send
shift up and shift down.